refactor: split API file to corresponding modules
