+++ src/modules/addressesList/api.js
@@ -1,0 +1,8 @@
+export const fetchAddressesList = () => {
+  return fetch("https://loft-taxi.glitch.me/addressList", {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json;charset=utf-8",
+    },
+  }).then((response) => response.json());
+};
