+++ src/modules/auth/api.js
@@ -1,0 +1,27 @@
+export const fetchLogin = ({ email, password }) => {
+  return fetch("https://loft-taxi.glitch.me/auth", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json;charset=utf-8",
+    },
+    body: JSON.stringify({
+      email,
+      password,
+    }),
+  }).then((response) => response.json());
+};
+
+export const fetchRegistration = ({ email, password, name, surname }) => {
+  return fetch("https://loft-taxi.glitch.me/register", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json;charset=utf-8",
+    },
+    body: JSON.stringify({
+      email,
+      password,
+      name,
+      surname,
+    }),
+  }).then((response) => response.json());
+};
